+++ src/lib/mino/data.ts
@@ -1,6 +1,7 @@
 import { range } from "lodash-es"
 import Vector, { type VectorLike } from "../vector"
 import type Polyomino from "./Polyomino"
+import PointSet from "$lib/PointSet"
 
 export type Dims = [number, number]
 export type Coord = Vector
@@ -84,6 +85,59 @@ export function hasX(mino: MinoData, x: number) {
 
 export function hasY(mino: MinoData, y: number) {
   return mino.values().some((v) => my(v) === y)
+}
+
+export function addSquare(mino: MinoData, [x, y]: VectorLike) {
+  if (x < 0) {
+    const result = new Set(mino.values().map((m) => m + mask(1, 0)))
+    result.add(mask(0, y))
+    return result
+  } else if (y < 0) {
+    const result = new Set(mino.values().map((m) => m + mask(0, 1)))
+    result.add(mask(x, 0))
+    return result
+  } else {
+    const result = new Set(mino)
+    result.add(mask(x, y))
+    return result
+  }
+}
+
+export function removeSquare(mino: MinoData, [x, y]: VectorLike) {
+  const clone = new Set(mino)
+  clone.delete(mask(x, y))
+  if (!hasX(clone, 0)) {
+    return new Set(clone.values().map((m) => m - mask(1, 0)))
+  }
+  if (!hasY(clone, 0)) {
+    return new Set(clone.values().map((m) => m - mask(0, 1)))
+  }
+  return clone
+}
+
+export function isValid(mino: MinoData): boolean {
+  const p0 = mino.values().next().value!
+  // the null-omino is not a valid polyomino
+  if (!p0) return false
+  const queue = [p0]
+
+  // Initialize the visited bitmask
+  // Include the mino's width so that we can easily compare later
+  let visited = new PointSet()
+
+  while (queue.length) {
+    const p = queue.pop()!
+    const v = Vector.of(unmask(p))
+    if (visited.has(v)) continue
+    visited.add(v)
+
+    for (const nbr of getNeighbors(Vector.fromArray(unmask(p)))) {
+      if (!mino.has(maskVec(nbr))) continue
+      queue.push(maskVec(nbr))
+    }
+  }
+  // True if we have visited all the squares in the mino
+  return visited.size === mino.size
 }
 
 /**
